Extract profile caching logic to a new credentialsGenerator
